pyproject.toml
[tool.poetry]
name = "competition-programs"
version = "0.1.0"
description = "The code OTTWO runs in competitions"
authors = ["GSG-Robots", "J0J0HA aka. JoJoJux", "Meganton"]
readme = "README.md"
packages = [{include = "*", from="src"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13.1,<3.14"
more-itertools = "^10.5.0"
spike-prime-connect = "^0.1.5"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"
sphinx-toolbox = "^3.8.1"
sphinx-autodoc-typehints = "^2.5.0"
sphinx-intl = "^2.3.1"
chardet = "^5.2.0"
sphinx-autobuild = "^2024.10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
